requirements.txt

aiogram~=3.7.0
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1